admin: text-align debug output dev: add user_id param to login event dev: no longer need hack for logout screen bugfix: reorder playlist broken with hidden tracks admin: update=1 now looks in local_files/lib too

+++ cchost_lib/cc-query.php
@@ -1383,8 +1383,9 @@ class CCQuery
 function cc_tcache_kill()
 {
     $files = glob(cc_temp_dir() . '/query_cache_*.txt');
-    foreach( $files as $file )
-        unlink($file);
+    if( $files !== false )
+        foreach( $files as $file )
+            unlink($file);
 }
